Atomic CVS commits: all or nothing.
To avoid half-exported HG revisions CVS commit procedure will lock
affected files and try commiting in dry-run mode. If it succeeds it
will then repeat commit in real.

WARNING:
In case dry-run commit succeeds and real commit fails (server went
down or comms broke) CVS repository will be left in undefined state
and manual intervention will be required to fix things.

To properly handle commits with multiple files that contain spaces
files.style has been introduced to tweak 'hg log' output and make it
parseable.

+++ hg2cvs.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+# set -x
+
 DEBUG=0
 DRY_RUN=0
 
@@ -30,34 +32,42 @@ do_cvs()
 
 add_cvsfolder()
 {
-    cvsflag=
-    if [ $DRY_RUN -eq 1 ]
-    then
-        cvsflag=-n
-    fi
-    a=$1
+    a="$1"
     p=
     first="/*"
     second="*/"
 
     if [ "$a" != "." ]
-    then 
+    then
         while [ "$a" != "$p" ]
         do
-            p=$a
-            f=${a/$first}
-            a=${a#$second}
-            cvs $cvsflag add $f
-            cd $f
+            p="$a"
+            f="${a/$first}"
+            a="${a#$second}"
+            do_cvs -Q add "$f"
+            cd "$f"
         done
-    fi  
+    fi
+}
+
+cvs_lock_files()
+{
+    if [ "x$@" != "x" ] ; then
+        do_cvs -Q admin -l "$@"
+    fi
+}
+
+cvs_unlock_files()
+{
+    if [ "x$@" != "x" ] ; then
+        do_cvs -Q admin -u "$@"
+    fi
 }
 
 do_cvsimport()
 {
     local hg_rev="$1"
     local is_merge="$2"
-    local tags="$3"
 
     hg up -C $hg_rev
     if [ $? -ne 0 ] ; then
@@ -65,51 +75,68 @@ do_cvsimport()
         return 1
     fi
 
-    added_files=$(hg log --template "{file_adds}\n" -r "$hg_rev" | perl -ne "s/ /\\ /g;print;")
+    local file_list=/tmp/hg2cvs.${hg_rev}.changes
+    hg log --style "$HG_FILES_STYLE" -r "$hg_rev" > $file_list
     if [ $? -ne 0 ] ; then
-        warning "Unable to identify added files"
+        warning "Unable to identify affected files"
+        rm -f $file_list
         return 1
     fi
 
-    changed_files=$(hg log --template "{file_mods}\n" -r "$hg_rev" | perl -ne "s/ /\\ /g;print;")
-    if [ $? -ne 0 ] ; then
-        warning "Unable to identify modified files"
-        return 1
+    local added_files=()
+    local changed_files=()
+    local removed_files=()
+
+    while read line
+    do
+        local mod_type=${line:0:1}
+        local mod_file="${line:2}"
+
+        # skip .hg* files
+        if [ "${mod_file:0:3}" = ".hg" ] ; then
+            continue;
+        fi
+
+        case $mod_type in
+            A) added_files[${#added_files[@]}]="$mod_file"     ;;
+            M) changed_files[${#changed_files[@]}]="$mod_file" ;;
+            R) deleted_files[${#deleted_files[@]}]="$mod_file" ;;
+            *) warning "Bad modification type $mod_type" ; return 1 ;;
+        esac
+
+    done < $file_list
+    rm $file_list
+
+    debug added: "${added_files[@]}"
+    debug changed: "${changed_files[@]}"
+    debug deleted: "${deleted_files[@]}"
+
+    local n_files=$((${#deleted_files[@]} + ${#added_files[@]} + ${#changed_files[@]}))
+    if [ "$n_files" -eq 0 ] ; then
+        debug "Nothing to commit to CVS"
+        return 0;
     fi
 
-    removed_files=$(hg log --template "{file_dels}\n" -r "$hg_rev" | perl -ne "s/ /\\ /g;print;")
-    if [ $? -ne 0 ] ; then
-        warning "Unable to identify removed files"
-        return 1
+    if [ ${#deleted_files[@]} -gt 0 ] ; then
+        do_cvs remove -f "${deleted_files[@]}"
     fi
 
-    added_files=${added_files/.hgtags/}
-    changed_files=${changed_files/.hgtags/}
+    for added in "${added_files[@]}"
+    do
+         pushd .
+         local dirpath=$(dirname "$added")
+         add_cvsfolder "$dirpath"
+         if [ $? -ne 0 ]; then
+             return 1
+         fi
 
-    debug "added: $added_files"
-    debug "changed: $changed_files"
-    debug "removed: $removed_files"
-
-    if [ -n "$added_files" ] ; then
-        for i in $added_files
-        do
-            pushd .
-            add_cvsfolder $(dirname $i)
-            if [ $? -ne 0 ]; then
-                return 1
-            fi
-
-            do_cvs add $(basename $i)
-            if [ $? -ne 0 ]; then
-                return 1
-            fi
-            popd
-        done
-    fi
-
-    if [ -n "$removed_files" ] ; then
-        do_cvs remove -f $removed_files
-    fi
+         local filename=$(basename "$added")
+         do_cvs add "$filename"
+         if [ $? -ne 0 ]; then
+             return 1
+         fi
+         popd
+    done
 
     local descfile=/tmp/hg2cvs.desc.$hg_rev
     local revspec=$hg_rev
@@ -118,9 +145,31 @@ do_cvsimport()
     fi
 
     hg log --template "{desc}\n-- {author} {date|isodatesec} commit {node|short}\n\n" -r "$revspec" > $descfile
-    do_cvs -Q commit -F $descfile $added_files $deleted_files $changed_files
+
+    cvs_lock_files "${deleted_files[@]}" "${changed_files[@]}"
+    if [ $? -ne 0 ] ; then
+        warning Unable to CVS-lock files
+        cvs_unlock_files "${deleted_files[@]}" "${changed_files[@]}"
+        return 1
+    fi
+
+    # dry run to check if commit succeeds
+    do_cvs -Q -n commit -F $descfile "${added_files[@]}" "${deleted_files[@]}" "${changed_files[@]}"
+    if [ $? -ne 0 ] ; then
+        warning CVS commit verification failed
+        cvs_unlock_files "${deleted_files[@]}" "${changed_files[@]}"
+        return 1
+    fi
+
+    # real commit, we don't expect it to fail
+    do_cvs -Q commit -F $descfile "${added_files[@]}" "${deleted_files[@]}" "${changed_files[@]}"
     local ret=$?
+    if [ $ret -ne 0 ] ; then
+        warning CVS commit failed, revision $hg_rev might be half-commited to CVS
+    fi
+
     rm $descfile
+    cvs_unlock_files "${deleted_files[@]}" "${changed_files[@]}"
 
     return $ret
 }
@@ -166,7 +215,7 @@ export_commits()
             echo "$last_imported_rev is a MERGE changeset";
         fi
 
-        do_cvsimport "$last_imported_rev" "$is_merge" "$tags"
+        do_cvsimport "$last_imported_rev" "$is_merge"
         if [ $? -ne 0 ] ; then
             warning "CVS import of $last_imported_rev failed"
             return 1
@@ -193,6 +242,7 @@ export_commits()
 ###############################################################################
 
 CVS_SANDBOX="$1"
+HG_FILES_STYLE=$(dirname $0)/files.style
 
 hg_branches=$(hg branches | cut -f 1 -d ' ')
